Feat/pp 18 implement software projects page (#4)
* re(home):move static layers to layout page
this commit moves Title, NavSidebar, and Footer components to layout.tsx to allow for static components

* feat(title):route title to home when clicked
this commit leverages useRouter from next/navigation and declares the title component as a client side rendering component to leverage useRouter capabilities

* re(home):adjust styling and font weights of main text
this commit makes the following stylistic changes:
- 'software engineer' and 'artist' is now displayed as 'software engineer!' and 'artist!'
- resume button is now listed as 'experiences'
- font weight of main text is thinner

* feat(software-projects):create Project interface and describe projects to display
this commit creates a Project interface which describes the data that makes up a project to be displayed.
this commit also fills out the projects to be displayed.:

* feat(software projects):create component to display software projects

* re(layout):change route to software projects to be more descriptive

* re(software projects):re order software projects for better work representation

* fix(nav sidebar)add method to handle routing to new pages

* feat(software projects):create image modal components and render dynamically
this commit creates ImageModal component and renders modal depending on if modal is toggled open

* feat(public):add gif to represent portfolio

* feat(software projects):create image modal
this commit ensures that the image modal is readable by changing its styling to be layered and with background color.
this commit also moves the ImageModal component to the components section so it can be reused for other sections

* re(navsidebar):adjust styling to make clearer what page site is on

* re(layout):remove instances of redux
this commit removes the instance of StoreProvider wrapper on layout page as it is no longer being leveraged by the rest of the application

* feat(software projects):create more info modal
this commit adds logic to the MoreInfoModal to display necessary data for each software project
state management was added to the page.tsx of software-projects directory in order to track the info modal

* feat(image modal):enable opening of info modal
this commit passes down the handleOpenText function which handles setting an opened state for the associated text modal

* feat(css):ensure that styling aligns with figma and best practices

---------

Co-authored-by: Kristinne Fragata <krisfragata@Kristinnes-MacBook-Air.local>