Feat/pp 18 implement software projects page 2 (#6)
* re(home):move static layers to layout page
this commit moves Title, NavSidebar, and Footer components to layout.tsx to allow for static components

* feat(title):route title to home when clicked
this commit leverages useRouter from next/navigation and declares the title component as a client side rendering component to leverage useRouter capabilities

* re(home):adjust styling and font weights of main text
this commit makes the following stylistic changes:
- 'software engineer' and 'artist' is now displayed as 'software engineer!' and 'artist!'
- resume button is now listed as 'experiences'
- font weight of main text is thinner

* feat(software-projects):create Project interface and describe projects to display
this commit creates a Project interface which describes the data that makes up a project to be displayed.
this commit also fills out the projects to be displayed.:

* feat(software projects):create component to display software projects

* re(layout):change route to software projects to be more descriptive

* re(software projects):re order software projects for better work representation

* fix(nav sidebar)add method to handle routing to new pages

* feat(software projects):create image modal components and render dynamically
this commit creates ImageModal component and renders modal depending on if modal is toggled open

* feat(public):add gif to represent portfolio

* feat(software projects):create image modal
this commit ensures that the image modal is readable by changing its styling to be layered and with background color.
this commit also moves the ImageModal component to the components section so it can be reused for other sections

* re(navsidebar):adjust styling to make clearer what page site is on

* re(layout):remove instances of redux
this commit removes the instance of StoreProvider wrapper on layout page as it is no longer being leveraged by the rest of the application

* feat(software projects):create more info modal
this commit adds logic to the MoreInfoModal to display necessary data for each software project
state management was added to the page.tsx of software-projects directory in order to track the info modal

* feat(image modal):enable opening of info modal
this commit passes down the handleOpenText function which handles setting an opened state for the associated text modal

* feat(css):ensure that styling aligns with figma and best practices

* re(css):adjust styling of info text
this commit ensures that text for each software project is visible and not cut off.
this commit also ensures that the images are visible and not cut off. this was done by adjusting margins

* re(software projects):toggle open and close of project when clicked
this commit makes changes to the list components by tweaking its onClick logic. Instead of setting state visibility to true, it is instead using toggle logic

* feat(image modal):add logic to display images
this commit adds logic to display images by adding the following
- renderImages method to return Image component if imageSources is present
- handleNextImage and handleBackImage handles moving through images on screen

* feat(tailwind):add color configurations

* feat(public):add reference images for software projects

* feat(software projects):ensure that text info opened is sent to front at all times
this commit adds logic to ensure that info modals opened by the 'more info..' button are sent to the front. This is done by dynamically adjusting the z-index of each modal when clicked. new methods were added to achieve this:
- state has been instantiated to track and set the 'z-index' of each info modal
- sendInfoToFront loops through all the z-index state setters and sets its value to either '' or 'z-100' depending on if it is recently clicked.

* fix(software projects):ensure that state is set to correct type
this method ensures that the sendInfoToFront method changes the state of z-index trackers to an empty string, '', instead of null to align with type definitions of each state

* re(css):adjust border of modals
this commit changes the border thickness of info modals to be thinner

* re(image modal):change the name of imported prop for better indication of function
this commit changes the passed in method 'handleOpenText' to 'handleTextToggle' as this better describes the indication of function

* fix(image modal):fix compiling error in path component
this commit changes the 'stroke-linecap' property to 'strokeLinecap' to align with next.js compilation expectations

* re(software projects):ensure that opened projects open on top of others (#5)

Co-authored-by: Kristinne Fragata <krisfragata@Kristinnes-MacBook-Air.local>

* feat(software projects):import photos for projects
this commit imports photos for the 'svisualize' project into the project list and adds them to its respective imageSources array
this commit also comments out the 'whats the flow' project as it is currently in progress

* re(image modal):adjust styling to display action buttons when necessary
this commit ensures that image action buttons are only visible when tehre are more than one image source
this commit also displays the amount of images shown

* re(css):adjust styling for image modals to fit index

* feat(public):import svisualize images to public directory

* fix(image modal):address compile errors in path component
rename css properties in path component to match with next's expectations. Such as, strokeLinecap, strokeLinejoin and strokeWidth instead of stroke-linecap, stroke-linejoin and stroke-width

* re(software projects):track software project metadata and states in an interface to avoid hardcoding components
this commit works to adhere to DRY and modularizes component rendering methods, avoiding hardcoding methods and avoids repeated switch statements
the following changes have been made:
- new ProjectControlData interface to track project metadata and state/setter methods
- projectControlData is defined to assign all displayed projects their states and setters. this array is now leveraged by all render methods such ast list, renderImageModals, renderInfoModals, and is also leveraged by all dynamic styling methods such as sendInfoToFront and sendImageToFront

* feat(public):import daily dose image sources

* re(software projects):remove unused imports and useEffect

---------

Co-authored-by: Kristinne Fragata <krisfragata@Kristinnes-MacBook-Air.local>
Co-authored-by: Kristinne Fragata <krisfragata@mac.lan>